Fix error in adding up yearly results for scenario
For a given simulation year `i` and eco factor index `j`, `yearTotals[i][j]` is supposed to sum values for that eco factor for all simulated trees. But as written it was replacing rather than summing, and so was just giving the value for the final tree. Now it sums correctly.